requirements.txt
bs4~=0.0.1
beautifulsoup4~=4.9.3
openpyxl~=3.0.5
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
